requirements.txt
streamlit
Pillow

torch
transformers
pandas
groq